UI: Move base ui classes into '.ui.base' package
