+++ app/src/main/java/soup/movie/ui/base/BaseViewModel.kt
@@ -1,8 +1,9 @@
-package soup.movie.ui
+package soup.movie.ui.base
 
 import androidx.lifecycle.ViewModel
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
+import soup.movie.ui.ResultMapper
 
 abstract class BaseViewModel : ViewModel(), ResultMapper {
