+++ app/src/main/java/soup/movie/ui/home/HomeFragment.kt
@@ -19,7 +19,7 @@ import soup.movie.R
 import soup.movie.analytics.EventAnalytics
 import soup.movie.data.MovieSelectManager
 import soup.movie.databinding.HomeFragmentBinding
-import soup.movie.ui.BaseFragment
+import soup.movie.ui.base.BaseFragment
 import soup.movie.util.consume
 import soup.movie.util.getColorAttr
 import soup.movie.util.lazyFast
