+++ app/src/main/java/soup/movie/ui/search/SearchFragment.kt
@@ -20,7 +20,7 @@ import soup.movie.R
 import soup.movie.analytics.EventAnalytics
 import soup.movie.data.MovieSelectManager
 import soup.movie.databinding.SearchFragmentBinding
-import soup.movie.ui.BaseFragment
+import soup.movie.ui.base.BaseFragment
 import soup.movie.ui.home.HomeListAdapter
 import soup.movie.util.ImeUtil
 import soup.movie.util.lazyFast
