+++ app/src/main/java/soup/movie/ui/theater/edit/TheaterEditViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import io.reactivex.android.schedulers.AndroidSchedulers
 import soup.movie.data.model.Theater
 import soup.movie.domain.theater.edit.TheaterEditManager
-import soup.movie.ui.BaseViewModel
+import soup.movie.ui.base.BaseViewModel
 import javax.inject.Inject
 
 class TheaterEditViewModel @Inject constructor(
