+++ app/src/main/java/soup/movie/ui/theme/ThemeOptionViewModel.kt
@@ -3,7 +3,7 @@ package soup.movie.ui.theme
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import soup.movie.theme.ThemeOptionManager
-import soup.movie.ui.BaseViewModel
+import soup.movie.ui.base.BaseViewModel
 import javax.inject.Inject
 
 class ThemeOptionViewModel @Inject constructor(
